added author to book

+++ dotnet_tasks/BookStore/WebApi/Application/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Linq;
+using Microsoft.EntityFrameworkCore;
 
 namespace WebApi.Application.AuthorOperations.Commands.DeleteAuthor;
 
@@ -15,9 +16,11 @@ public class DeleteAuthorCommand
 
     public void Handle()
     {
-        var author = context.Authors.SingleOrDefault(x => x.Id == AuthorId);
+        var author = context.Authors.Include(a => a.Books).SingleOrDefault(x => x.Id == AuthorId);
         if (author is null)
             throw new InvalidOperationException("The author to be deleted was not found");
+        if (author.Books.Any())
+            throw new InvalidOperationException("Author have book(s) so you can't delete");
 
         context.Authors.Remove(author);
         context.SaveChanges();
